Refactor enum dependencies between schema and models.
This is a piece that is needed to create a schema package - which I would really like to have around for #15639 and to in general write higher-level API tests.
